feat: update

+++ src/services/players-service.ts
@@ -1,4 +1,5 @@
 import { PlayerModel } from "../models/players-model";
+import { StatisticsModel } from "../models/statistics-model";
 import * as PlayersRepository from "../repositories/players-repository";
 import * as HttpResponse from "../utils/http.helper";
 
@@ -36,9 +37,9 @@ export const createPlayerService = async (player: PlayerModel) => {
   if (Object.keys(player).length !== 0) {
     await PlayersRepository.insertPlayer(player);
 
-    response = HttpResponse.created();
+    response = await HttpResponse.created();
   } else {
-    response = HttpResponse.badRequest();
+    response = await HttpResponse.badRequest();
   }
 
   return response;
@@ -47,9 +48,31 @@ export const createPlayerService = async (player: PlayerModel) => {
 export const deletePlayerService = async (id: number) => {
   let response = null;
 
-  await PlayersRepository.deleteOnePlayer(id);
+  const data = PlayersRepository.findPlayerById(id);
 
-  response = HttpResponse.ok({ message: "deleted" });
+  if (Object.keys(data).length === 0) {
+    response = await HttpResponse.badRequest();
+  } else {
+    await PlayersRepository.deleteOnePlayer(id);
+    response = await HttpResponse.ok({ message: "deleted" });
+  }
+
+  return response;
+};
+
+export const updatePlayerService = async (
+  id: number,
+  statistics: StatisticsModel
+) => {
+  let response = null;
+
+  const data = await PlayersRepository.findAndModifyPlayer(id, statistics);
+
+  if (Object.keys(data).length === 0) {
+    response = await HttpResponse.badRequest();
+  }
+
+  response = await HttpResponse.ok(data);
 
   return response;
 };
